Logged out form and form styles and functionality fixes
