+++ website/src/app/dashboard/jobs/[id]/JobChatsList.tsx
@@ -14,13 +14,6 @@ const JobChatsList = ({users, job, setSelectedWorker} : {users: Record<string, U
             <span className='text-darkBlueFont text-sm'>{numberOfWorkers}/{numberOfWorkers} applicants reviewed</span>  
         </div>
         <ul>
-            <li key={'first'} className='flex flex-row py-2 hover:bg-slate-100 cursor-pointer rounded' onClick={() => setSelectedWorker('allEvents')}>
-                <Image className='flex flex-1 items-center mr-2 max-h-10 max-w-10 rounded-lg' src={'/profilePicture.webp'} height={100} width={100} alt={'Profile picture'}></Image>
-                <div className='flex flex-col flex-[4] items-left'>
-                    <span>Global Events</span>
-                    <p className='text-sm text-darkBlueFont truncate ... max-w-56 '>Everything Everywhere All at Once</p>
-                </div>
-            </li>
             {  Object.entries(users).map(([key, value]) => (
                 job?.roles.creator !== key ?
                     <li key={key} className='flex flex-row py-2 hover:bg-slate-100 cursor-pointer rounded' onClick={() => setSelectedWorker(key)}>
