+++ website/src/components/Textarea.tsx
@@ -3,18 +3,17 @@ import { clsx } from 'clsx'
 import { forwardRef } from 'react'
 
 export const Textarea = forwardRef<HTMLTextAreaElement, { resizable?: boolean } & HeadlessTextareaProps>(
-  function Textarea({ className, resizable = true, ...props }, ref) {
+  function Textarea({ className, resizable = true,...props }, ref) {
     return (
       <span
         data-slot="control"
         className={clsx([
           className,
-
           // Basic layout
-          'relative block w-full !mt-1.1',
+          'relative block w-full',
 
           // Background color + shadow applied to inset pseudo element, so shadow blends with border in light mode
-          'before:absolute before:inset-px before:rounded-[calc(theme(borderRadius.xl)-1px)] before:bg-white before:shadow',
+          'before:absolute before:inset-px before:bg-white before:shadow',
 
           // Background color is moved to control and shadow is removed in dark mode so hide `before` pseudo
           'dark:before:hidden',
@@ -29,8 +28,9 @@ export const Textarea = forwardRef<HTMLTextAreaElement, { resizable?: boolean } 
         <HeadlessTextarea
           ref={ref}
           className={clsx([
+            className,
             // Basic layout
-            'relative block h-full w-full appearance-none rounded-xl px-[calc(theme(spacing[3.5])-1px)] py-[calc(theme(spacing[2.5])-1px)] sm:px-[calc(theme(spacing.3)-1px)] sm:py-[calc(theme(spacing[1.5])-1px)]',
+            'relative block h-full w-full appearance-none rounded-xl px-[calc(theme(spacing[3.5])-1px)] py-[calc(theme(spacing[2.5])-1px)] sm:px-[calc(theme(spacing.3)-1px)] sm:py-[calc(theme(spacing[1.5])-1px)] align-center',
 
             // Typography
             'text-base/6 text-zinc-950 placeholder:text-zinc-500 sm:text-sm/6 dark:text-white',
@@ -42,7 +42,7 @@ export const Textarea = forwardRef<HTMLTextAreaElement, { resizable?: boolean } 
             'bg-transparent dark:bg-white/5',
 
             // Hide default focus styles
-            'focus:outline-none focus:ring-1 focus:ring-primary focus:border-0',
+            'focus:outline-none focus:!ring-0 focus:border-primary',
 
             // Invalid state
             'data-[invalid]:border-red-500 data-[invalid]:data-[hover]:border-red-500 data-[invalid]:dark:border-red-600 data-[invalid]:data-[hover]:dark:border-red-600',
